Navbar functionality and styling done
