+++ frontend/package.json
@@ -12,7 +12,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router": "^6.3.0",
-    "react-router-dom": "^6.3.0",
+    "react-router-dom": "^6.9.0",
     "react-scripts": "^5.0.1",
     "tailwindcss": "^3.2.7",
     "web-vitals": "^2.1.4"
