+++ frontend/src/components/feed/Feed.js
@@ -1,12 +1,9 @@
-import React, { useEffect, useState, useContext } from "react";
+import React, { useEffect, useState } from "react";
 import PropTypes from "prop-types";
 import Post from "../post/Post";
 import NewPost from "../newPost/NewPost";
-import Button from "../button/Button";
-import { ModalContext } from "../../contexts/modalContext";
 
 const Feed = ({ navigate }) => {
-  const { pushModal } = useContext(ModalContext);
   const [posts, setPosts] = useState([]);
   const [token, setToken] = useState(window.localStorage.getItem("token"));
 
@@ -33,28 +30,9 @@ const Feed = ({ navigate }) => {
     getPosts();
   }, []);
 
-  const logout = () => {
-    window.localStorage.removeItem("token");
-    pushModal({
-      message: "Successfully logged out",
-      type: "success",
-    });
-    navigate("/login");
-  };
-
   if (token) {
     return (
       <>
-        <Button
-          text="Logout"
-          clickCallback={logout}
-          type="button"
-          id="logout"
-          className="max-w-sm"
-        />
-
-        <h2>Posts</h2>
-
         <NewPost getPosts={getPosts} />
 
         <div id="feed" role="feed">
